fastapi_app/requirements.txt
fastapi
httpx
aiofiles
pydantic
motor
pymongo
uvicorn
django-environ
redis
python-multipart
django
pymongo